Cart_quantity_manage Order_History User_update User_Pass_Change

+++ urbanbazaarapp/views.py
@@ -1,12 +1,75 @@
 from django.views import View
 from django.shortcuts import render , redirect
-from .forms import SignupForm , ProductForm
-from .models import Profile , User , Cart , CartItem , Product , Order , OrderItem
-from django.contrib.auth import authenticate , login as auth_login , logout
+from .forms import SignupForm , ProductForm , ProductUpdateForm , User_update_form
+from .models import Profile , User , Cart , CartItem , Product , Order , OrderItem 
+from django.contrib.auth import authenticate , login as auth_login , logout , update_session_auth_hash
 from django.core.paginator import Paginator
 from django.contrib import messages
 from datetime import datetime
 from django.shortcuts import render, redirect, get_object_or_404
+from django.contrib.auth.forms import PasswordChangeForm
+
+
+class SignupUser(View):  
+    def get(self, request): 
+        if not request.user.is_authenticated:
+            signup_form = SignupForm()
+            return render(request, 'signup.html', {'form': signup_form})
+        return redirect('login')
+        
+    def post(self, request):  
+        if not request.user.is_authenticated:
+            signup_form = SignupForm(data = request.POST)
+            if signup_form.is_valid():
+                user = signup_form.save()
+                role = signup_form.cleaned_data['role']
+                Profile.objects.create(user = user, role = role)
+                user.save()
+                return redirect('login')
+            return render(request, 'signup.html', {'form': signup_form})
+        return redirect('login')
+
+
+class User_Pass_Change(View):
+    def get(self, request ,  *args, **kwargs): 
+        user_id = self.kwargs.get('id')
+        if request.user.is_authenticated and request.user.id == user_id :
+            user_form = PasswordChangeForm(user = request.user )
+            return render(request, 'changepass.html', {'form': user_form})
+        
+        return redirect('login')
+
+    def post(self, request ,  *args, **kwargs):  
+        user_id = self.kwargs.get('id')
+        if request.user.is_authenticated and request.user.id == user_id :
+            user_form = PasswordChangeForm(data=request.POST, user = request.user )
+            if user_form.is_valid():
+                user_form.save()
+                update_session_auth_hash(request , user_form.user)
+                return redirect('login')
+        return redirect('login')
+
+
+class User_update(View):
+    def get(self, request ,  *args, **kwargs): 
+        user_id = self.kwargs.get('id')
+        if request.user.is_authenticated and request.user.id == user_id:
+            user = User.objects.get(id = user_id)
+            user_form = User_update_form(instance=user)
+            return render(request, 'User_update.html', {'form': user_form})
+        return redirect('login')
+
+    def post(self, request ,  *args, **kwargs):  
+        user_id = self.kwargs.get('id')
+        if request.user.is_authenticated and request.user.id == user_id:
+            user = User.objects.get(id=user_id)
+            user_form = User_update_form(data=request.POST, instance=user)
+            if user_form.is_valid():
+                user_form.save( )
+                user_form.save()
+                return redirect('redirect_view')
+            return redirect('User_update')
+        return redirect('login')
 
 
 class Userlogin(View):
@@ -39,45 +102,6 @@ class LogoutView(View):
         return redirect('login')
 
 
-class SignupUser(View):  
-    def get(self, request): 
-        if not request.user.is_authenticated:
-            signup_form = SignupForm()
-            return render(request, 'signup.html', {'form': signup_form})
-        return redirect('login')
-        
-    def post(self, request):  
-        if not request.user.is_authenticated:
-            signup_form = SignupForm(data = request.POST)
-            if signup_form.is_valid():
-                user = signup_form.save()
-                role = signup_form.cleaned_data['role']
-                Profile.objects.create(user = user, role = role)
-                user.save()
-                return redirect('login')
-            return render(request, 'signup.html', {'form': signup_form})
-        return redirect('login')
-
-
-class ProductAdd(View):  
-    def get(self, request): 
-        if request.user.is_authenticated and request.user.profile.role == 'Vendor':
-            product_form = ProductForm()
-            return render(request, 'product.html', {'form': product_form})
-        return redirect('login')
-
-    def post(self, request):  
-        if request.user.is_authenticated and request.user.profile.role == 'Vendor':
-            product_form = ProductForm(data = request.POST)
-            if product_form.is_valid():
-                user = product_form.save( commit = False)
-                user.vendor = request.user
-                user.save()
-                return redirect('Vendor_dashboard')
-            return render(request, 'product.html', {'form': product_form})
-        return redirect('login')
-
-
 class RedirectView(View):
     def get(self , request):
         if request.user.is_authenticated and request.user.is_superuser:
@@ -94,6 +118,9 @@ class Customer_dashboard(View):
     def get(self, request):
         if request.user.is_authenticated:
             try:
+                cart_items = CartItem.objects.filter(cart__user=request.user)
+                total_quantity = sum(item.quantity for item in cart_items)
+
                 profile = Profile.objects.get(user=request.user)
                 if profile.role == 'Customer':
                     category = request.GET.get('category')
@@ -108,7 +135,8 @@ class Customer_dashboard(View):
                     return render(request, 'Customer.html', {
                         'name': request.user.username,
                         'cart': cart,
-                        'products': products
+                        'products': products ,
+                        'total_quantity': total_quantity
                     })
                 else:
                     return redirect('login')
@@ -137,6 +165,48 @@ class Vendor_dashboard(View):
             except Profile.DoesNotExist:
                 return redirect('login')
             
+        return redirect('login')
+
+
+class ProductAdd(View):  
+    def get(self, request): 
+        if request.user.is_authenticated and request.user.profile.role == 'Vendor':
+            product_form = ProductForm()
+            return render(request, 'product.html', {'form': product_form})
+        return redirect('login')
+
+    def post(self, request):  
+        if request.user.is_authenticated and request.user.profile.role == 'Vendor':
+            product_form = ProductForm(data = request.POST , files=request.FILES)
+            if product_form.is_valid():
+                user = product_form.save( commit = False)
+                user.vendor = request.user
+                user.save()
+                return redirect('Vendor_dashboard')
+            return render(request, 'product.html', {'form': product_form})
+        return redirect('login')
+
+
+class ProductUpdate(View):
+    def get(self, request ,  *args, **kwargs): 
+        if request.user.is_authenticated and request.user.profile.role == 'Vendor':
+            product_id = self.kwargs.get('id')
+            product = Product.objects.get(id = product_id)
+            product_form = ProductUpdateForm(instance=product)
+            return render(request, 'update_product.html', {'form': product_form})
+        return redirect('login')
+
+    def post(self, request ,  *args, **kwargs):  
+        if request.user.is_authenticated and request.user.profile.role == 'Vendor':
+            product_id = self.kwargs.get('id')
+            product = Product.objects.get(id=product_id)
+            product_form = ProductUpdateForm(data=request.POST, instance=product , files=request.FILES)
+            if product_form.is_valid():
+                user = product_form.save( commit = False)
+                user.vendor = request.user
+                user.save()
+                return redirect('Vendor_dashboard')
+            return render(request, 'update_product.html', {'form': product_form})
         return redirect('login')
 
 
@@ -181,32 +251,27 @@ class Admin_dashboard(View):
 
 
 class Add_to_cart(View):
-    def post(self , request , *args, **kwargs):
+    def post(self, request, *args, **kwargs):
         if request.user.is_authenticated and request.user.profile.role == 'Customer':
             product_id = self.kwargs.get('id')
-            product = Product.objects.get(id = product_id)
-            cart , cart_created = Cart.objects.get_or_create(user = request.user)
-            # cart = Cart.objects.get(user = request.user)
-            cart_item , item_created = CartItem.objects.get_or_create(cart = cart , product = product , defaults={'quantity' : 1})
+            product = Product.objects.get(id=product_id)
+            cart, cart_created = Cart.objects.get_or_create(user=request.user)
+            cart_item, item_created = CartItem.objects.get_or_create(cart=cart, product=product, defaults={'quantity': 1})
 
             if not item_created:
-                        if cart_item.quantity < product.stock_quantity:
-                            cart_item.quantity += 1
-                            cart_item.total_price = cart_item.product.price * cart_item.quantity 
-                            cart_item.save()
-                        else:
-                            if cart_item.quantity >= product.stock_quantity:
-                                messages.error(request, f"Only {product.stock_quantity} units available in stock.")
-                            else:
-                                pass
-                            
-                        cart_item.save()
-
-                        return redirect('Customer_dashboard')
+                if cart_item.quantity < product.stock_quantity:
+                    cart_item.quantity += 1
+                    cart_item.total_price = cart_item.product.price * cart_item.quantity
+                    cart_item.save()
+                    print(cart_item.quantity)
+                else:
+                    messages.error(request, f"Only {product.stock_quantity} units available in stock.")
+                
+                return redirect('Customer_dashboard')
             else:
-                return redirect('login')
-            
-        return redirect('login')
+                return redirect('Customer_dashboard')
+        else:
+            return redirect('login')
 
 
 class View_cart(View):
@@ -251,8 +316,6 @@ class Checkout(View):
                 cart = Cart.objects.filter(user=request.user).first()
                 cart_items = CartItem.objects.filter(cart=cart)
 
-                total_amount = sum(item.product.price * item.quantity for item in cart_items)
-
                 for item in cart_items:
                     item.total_price = item.product.price * item.quantity
 
@@ -262,12 +325,13 @@ class Checkout(View):
                     OrderItem.objects.create(order=order,product=item.product,quantity=item.quantity,price=item.product.price*item.quantity)
                     item.product.stock_quantity -= item.quantity
                     item.product.save()
-                    cart_items.delete()
-                    return redirect('Customer_dashboard')
+                cart_items.delete()
 
-                
-                return render(request, 'checkout.html', {'cart_items': cart_items, 'total_amount': total_amount})
+                messages.success(request, "Your order has been successfully placed!")
+                return redirect('Customer_dashboard')  
+
         else:
+            messages.error(request, "You need to log in to proceed with the checkout.")
             return redirect('login')
 
 
@@ -290,23 +354,53 @@ class Update_Order_Status(View):
             order.save()
             return redirect('Admin_dashboard')
 
-        return render(request, 'Update_Order_Status.html', {'order': order})
-
 
 class Cart_quantity_manage(View):
     def post(self, request, *args, **kwargs):
         if request.user.is_authenticated and request.user.profile.role == 'Customer':
             product_id = self.kwargs.get('id')
-            del_cart = CartItem.objects.get(id = product_id)
-            if del_cart.quantity >1:
-                del_cart.quantity -= 1
-                del_cart.total_price = del_cart.product.price * del_cart.quantity
-                del_cart.save()
-            else:
-                del_cart.delete()
-            return redirect('View_cart')
+            del_cart = CartItem.objects.get(id=product_id)
+            action = request.POST.get('action')
+            if action == 'decrease':
+                if del_cart.quantity > 1:
+                    del_cart.quantity -= 1
+                    del_cart.total_price = del_cart.product.price * del_cart.quantity
+                    del_cart.save()
+                else:
+                    del_cart.delete()
+                return redirect('View_cart')
+
+            elif action == 'increase':
+                if del_cart.quantity < del_cart.product.stock_quantity:
+                    del_cart.quantity += 1
+                    del_cart.total_price = del_cart.product.price * del_cart.quantity
+                    del_cart.save()
+                else:
+                    messages.error(request, f"Only {del_cart.product.stock_quantity} units available in stock.")
+                return redirect('View_cart')
+
+        return redirect('login')
+
+
+class Order_History(View):
+    def get(self, request ,  *args, **kwargs): 
+        if request.user.is_authenticated and request.user.profile.role == 'Customer':
+
+            orders = Order.objects.filter(customer = request.user)
+            order_items = OrderItem.objects.filter(order__in = orders)
+
+            order_paginator = Paginator(orders, 4) 
+            order_page_number = request.GET.get('order_page')
+            orders_page = order_paginator.get_page(order_page_number)
+            return render(request , 'order_history.html' , {
+                'orders':orders_page , 
+                'order_items':order_items
+
+                })
         
         return redirect('login')
+
+
